time precision increased, philos sleep until time passed since they change state, instead of sleep ttx time since last instruction

+++ philo_one/utils.h
@@ -4,7 +4,7 @@
 # include <stdlib.h>
 
 int			ft_is_digit(char c);
-uint64_t	ft_now(void);
+long long int	ft_now(void);
 int			ft_atoi(const char *str);
 int			ft_isint(const char *str);
 int			ft_strlen(const char *str);
